ipops: Move structures to header
